refactor(deployment): move deployment tasks
