+++ tasks/deploy-root-child.ts
@@ -10,7 +10,7 @@ import logger from "consola";
 import { HttpNetworkConfig } from "hardhat/src/types/config";
 import { deployToken } from "./utils/deploy-token";
 import { deployChildChainManager } from "./utils/deploy-child-chain-manager";
-import { RoleConstants } from "../common/constants";
+import { RoleConstants } from "../src/common/constants";
 
 task("deploy:complete")
   .setDescription("Deploys root token and its chain manager")
